airflow/requirements.txt
google-generativeai
openai
psycopg2-binary
sqlalchemy
